+++ bld/wgml1/c/gsga.c
@@ -185,8 +185,8 @@ static bool     get_att_name( const char *p, char *attname )
 
 
 static  char        stringval[VAL_LENGTH + 1];
-static  char    *   valptr;
-static  long        ranges[4];
+static  char        *valptr;
+static  int         ranges[4];
 
 
 /***************************************************************************/
@@ -315,8 +315,8 @@ static  condcode    scan_att_optionsB( gavalflags * val_flags, condcode cca,
             gn.arg.s = scan_start;
             gn.arg.e = scan_stop;
             gn.ignore_blanks = false;
-            ranges[2] = LONG_MIN;
-            ranges[3] = LONG_MIN;
+            ranges[2] = INT_MIN;
+            ranges[3] = INT_MIN;
             for( k = 0; k < 4; k++ ) {  // scan max 4 numbers
 
                 cc = getnum( &gn );
@@ -349,7 +349,7 @@ static  condcode    scan_att_optionsB( gavalflags * val_flags, condcode cca,
                     return( cc );
                 }
             }
-            if( ranges[2] > LONG_MIN ) {
+            if( ranges[2] > INT_MIN ) {
                 *val_flags |= val_def;  // we have default
                 *att_flags |= att_def;  // we have default
             }
@@ -377,7 +377,7 @@ static  condcode    scan_att_optionsB( gavalflags * val_flags, condcode cca,
                 return( cc );
             } else {
                 scan_start = gn.arg.s;
-                ranges [0] = gn.result;
+                ranges[0] = gn.result;
             }
         } else {
             cc = neg;
